# nodejs-server/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Recursos disponiveis para o aplicativo Smart POS"
  version: "1.0.0"
  title: "Linx Postos SmartPOS"
host: "postoshub.linx.com.br"
basePath: "/api"
schemes:
- "https"
paths:
  /v1/auth/partner/:
    post:
      tags:
      - "Auth"
      summary: "Login do parceiro"
      description: "Realiza o login do parceiro"
      operationId: "services.auth.web.post_auth_partner"
      parameters:
      - name: "Store-CNPJ"
        in: "header"
        description: "CNPJ da empresa (somente digitos)"
        required: true
        type: "string"
      - name: "APK-Version"
        in: "header"
        description: "Número da versão do aplicativo"
        required: false
        type: "string"
      - name: "Device-Number"
        in: "header"
        description: "Número serial do dispositivo"
        required: true
        type: "string"
      - name: "RFID-Available"
        in: "header"
        description: "Indica se o dispositivo faz leitura de cartão RFID"
        required: false
        type: "boolean"
      - in: "body"
        name: "data"
        description: "Dados do parceiro"
        required: false
        schema:
          $ref: "#/definitions/data"
      responses:
        "200":
          description: "Login foi bem sucedido"
          schema:
            $ref: "#/definitions/inline_response_200"
        "400":
          description: "Requisição com sintaxe inválida"
          schema:
            $ref: "#/definitions/Error"
        "401":
          description: "Recurso solicitado requer autenticação"
          schema:
            $ref: "#/definitions/Error"
        "403":
          description: "Sem permissão para acesso ao conteúdo"
          schema:
            $ref: "#/definitions/Error"
        "404":
          description: "Recurso solicitado não encontrado"
          schema:
            $ref: "#/definitions/Error"
        "405":
          description: "Método da requisição não permitido"
          schema:
            $ref: "#/definitions/Error"
        "409":
          description: "Existe uma sessão já iniciada ou conflitante com a tentativa\
            \ corrente"
          schema:
            $ref: "#/definitions/Error"
        "415":
          description: "Formato do conteúdo de dados não suportado"
          schema:
            $ref: "#/definitions/Error"
        "500":
          description: "Erro interno no servidor"
          schema:
            $ref: "#/definitions/Error"
      x-swagger-router-controller: "Auth"
...
  /v1/customer/:
    get:
      tags:
      - "Customer"
      summary: "Consulta de clientes"
      description: "Realiza a consulta de clientes"
      operationId: "services.customer.web.get_customer"
      parameters:
      - name: "Authorization"
        in: "header"
        description: "Token de autorização da API gerado pelo hub de serviços (Em\
          \ rede local gerado pelo backoffice)"
        required: true
        type: "string"
      - name: "POS-Token"
        in: "header"
        description: "Token de autenticação gerado pelo backoffice"
        required: true
        type: "string"
      - name: "Store-CNPJ"
        in: "header"
        description: "CNPJ da empresa (somente digitos)"
        required: true
        type: "string"
      - name: "APK-Version"
        in: "header"
        description: "Número da versão do aplicativo"
        required: false
        type: "string"
      - name: "Device-Number"
        in: "header"
        description: "Número serial do dispositivo"
        required: true
        type: "string"
      - name: "RFID-Available"
        in: "header"
        description: "Indica se o dispositivo faz leitura de cartão RFID"
        required: false
        type: "boolean"
      - name: "search"
        in: "query"
        description: "Valor de pesquisa (ID, código, nome, apelido, CPF, CNPJ, RFID\
          \ do cliente, placa do veículo, ID ou CPF do motorista, cartão fidelidade)"
        required: false
        type: "string"
      - name: "field"
        in: "query"
        description: "Campo de pesquisa (id, code, name, nickname, cpf, rfid, plate,\
          \ loyaltyCard)"
        required: false
        type: "string"
      - name: "page"
        in: "query"
        description: "Número da página (Nulo para retornar todos registros)"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "Limite de registros (Nulo para retornar todos registros)"
        required: false
        type: "integer"
      responses:
        "200":
          description: "Consulta de clientes foi bem sucedida"
          schema:
            $ref: "#/definitions/inline_response_200_7"
      x-swagger-router-controller: "Customer"
  /v1/customer/credit/:
    get:
      tags:
      - "Customer"
      summary: "Consulta de restrição de crédito do cliente"
      description: "Realiza a consulta de restrição de crédito do cliente"
      operationId: "services.customer.web.get_customer_credit"
      parameters:
      - name: "Authorization"
        in: "header"
        description: "Token de autorização da API gerado pelo hub de serviços (Em\
          \ rede local gerado pelo backoffice)"
        required: true
        type: "string"
      - name: "POS-Token"
        in: "header"
        description: "Token de autenticação gerado pelo backoffice"
        required: true
        type: "string"
      - name: "Store-CNPJ"
        in: "header"
        description: "CNPJ da empresa (somente digitos)"
        required: true
        type: "string"
      - name: "APK-Version"
        in: "header"
        description: "Número da versão do aplicativo"
        required: false
        type: "string"
      - name: "Device-Number"
        in: "header"
        description: "Número serial do dispositivo"
        required: true
        type: "string"
      - name: "RFID-Available"
        in: "header"
        description: "Indica se o dispositivo faz leitura de cartão RFID"
        required: false
        type: "boolean"
      - name: "search"
        in: "query"
        description: "Valor de pesquisa (ID, código, nome, apelido, CPF, CNPJ do cliente,\
          \ placa do veículo)"
        required: false
        type: "string"
      - name: "paymentId"
        in: "query"
        description: "Id da forma de pagamento (Consulta a restrição de uma determinada\
          \ forma de pagamento)"
        required: false
        type: "integer"
      - name: "username"
        in: "query"
        description: "Nome do usuário com permissão para realizar a consulta de cré\
          dito completa"
        required: false
        type: "string"
      - name: "password"
        in: "query"
        description: "Senha de acesso do usuário com permissão para realizar a consulta\
          \ de crédito completa"
        required: false
        type: "string"
      - name: "rfid"
        in: "query"
        description: "Cartão RFID do usuário"
        required: false
        type: "string"
      responses:
        "200":
          description: "Constulta de restrição de crédito foi bem sucedida"
          schema:
            $ref: "#/definitions/inline_response_200_8"
      x-swagger-router-controller: "Customer"
  /v1/customer/{customerId}/driver/:
    get:
      tags:
      - "Customer"
      summary: "Consulta de motoristas do cliente"
      description: "Realiza a consulta de motoristas do cliente"
      operationId: "services.customer.web.get_customer_driver"
      parameters:
      - name: "Authorization"
        in: "header"
        description: "Token de autorização da API gerado pelo hub de serviços (Em\
          \ rede local gerado pelo backoffice)"
        required: true
        type: "string"
      - name: "POS-Token"
        in: "header"
        description: "Token de autenticação gerado pelo backoffice"
        required: true
        type: "string"
      - name: "Store-CNPJ"
        in: "header"
        description: "CNPJ da empresa (somente digitos)"
        required: true
        type: "string"
      - name: "APK-Version"
        in: "header"
        description: "Número da versão do aplicativo"
        required: false
        type: "string"
      - name: "Device-Number"
        in: "header"
        description: "Número serial do dispositivo"
        required: true
        type: "string"
      - name: "RFID-Available"
        in: "header"
        description: "Indica se o dispositivo faz leitura de cartão RFID"
        required: false
        type: "boolean"
      - name: "customerId"
        in: "path"
        description: "Id do cliente"
        required: true
        type: "integer"
      - name: "search"
        in: "query"
        description: "Pesquisa pelo nome do motorista"
        required: false
        type: "string"
      responses:
        "200":
          description: "Consulta de motoristas do cliente foi bem sucedida"
          schema:
            type: "array"
            description: "Lista de motoristas"
            items:
              $ref: "#/definitions/inline_response_200_9"
      x-swagger-router-controller: "Customer"
  /v1/customer/{customerId}/fleet/:
    get:
      tags:
      - "Customer"
      summary: "Consulta de veículos da frota do cliente"
      description: "Realiza a consulta de veículos da frota do cliente"
      operationId: "services.customer.web.get_customer_fleet"
      parameters:
      - name: "Authorization"
        in: "header"
        description: "Token de autorização da API gerado pelo hub de serviços (Em\
          \ rede local gerado pelo backoffice)"
        required: true
        type: "string"
      - name: "POS-Token"
        in: "header"
        description: "Token de autenticação gerado pelo backoffice"
        required: true
        type: "string"
      - name: "Store-CNPJ"
        in: "header"
        description: "CNPJ da empresa (somente digitos)"
        required: true
        type: "string"
      - name: "APK-Version"
        in: "header"
        description: "Número da versão do aplicativo"
        required: false
        type: "string"
      - name: "Device-Number"
        in: "header"
        description: "Número serial do dispositivo"
        required: true
        type: "string"
      - name: "RFID-Available"
        in: "header"
        description: "Indica se o dispositivo faz leitura de cartão RFID"
        required: false
        type: "boolean"
      - name: "customerId"
        in: "path"
        description: "Id do cliente"
        required: true
        type: "integer"
      - name: "search"
        in: "query"
        description: "Pesquisa pela placa do veículo"
        required: false
        type: "string"
      responses:
        "200":
          description: "Consulta de veículos da frota do cliente foi bem sucedida"
          schema:
            type: "array"
            description: "Lista de veículos da frota"
            items:
              $ref: "#/definitions/inline_response_200_10"
      x-swagger-router-controller: "Customer"
...
  /v1/employee/:
    get:
      tags:
      - "Employee"
      summary: "Consulta de funcionários"
      description: "Realiza a consulta de funcionários"
      operationId: "services.employee.web.get_employee"
      parameters:
      - name: "Authorization"
        in: "header"
        description: "Token de autorização da API gerado pelo hub de serviços (Em\
          \ rede local gerado pelo backoffice)"
        required: true
        type: "string"
      - name: "POS-Token"
        in: "header"
        description: "Token de autenticação gerado pelo backoffice"
        required: true
        type: "string"
      - name: "Store-CNPJ"
        in: "header"
        description: "CNPJ da empresa (somente digitos)"
        required: true
        type: "string"
      - name: "APK-Version"
        in: "header"
        description: "Número da versão do aplicativo"
        required: false
        type: "string"
      - name: "Device-Number"
        in: "header"
        description: "Número serial do dispositivo"
        required: true
        type: "string"
      - name: "RFID-Available"
        in: "header"
        description: "Indica se o dispositivo faz leitura de cartão RFID"
        required: false
        type: "boolean"
      - name: "search"
        in: "query"
        description: "Valor de pesquisa (ID, código, nome, apelido, CPF, RFID do funcioná\
          rio)"
        required: false
        type: "string"
      - name: "field"
        in: "query"
        description: "Campo de pesquisa (id, code, name, nickname, cpf, rfid)"
        required: false
        type: "string"
      - name: "page"
        in: "query"
        description: "Número da página (Nulo para retornar todos registros)"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "Limite de registros (Nulo para retornar todos registros)"
        required: false
        type: "integer"
      responses:
        "200":
          description: "Consulta de funcionários foi bem sucedida"
          schema:
            $ref: "#/definitions/inline_response_200_11"
      x-swagger-router-controller: "Employee"
...
  /v1/payment/:
    get:
      tags:
      - "Payment"
      summary: "Consulta das formas de pagamento"
      description: "Realiza a consulta das formas de pagamento"
      operationId: "services.payment.web.get_payment"
      parameters:
      - name: "Authorization"
        in: "header"
        description: "Token de autorização da API gerado pelo hub de serviços (Em\
          \ rede local gerado pelo backoffice)"
        required: true
        type: "string"
      - name: "POS-Token"
        in: "header"
        description: "Token de autenticação gerado pelo backoffice"
        required: true
        type: "string"
      - name: "Store-CNPJ"
        in: "header"
        description: "CNPJ da empresa (somente digitos)"
        required: true
        type: "string"
      - name: "APK-Version"
        in: "header"
        description: "Número da versão do aplicativo"
        required: false
        type: "string"
      - name: "Device-Number"
        in: "header"
        description: "Número serial do dispositivo"
        required: true
        type: "string"
      - name: "RFID-Available"
        in: "header"
        description: "Indica se o dispositivo faz leitura de cartão RFID"
        required: false
        type: "boolean"
      responses:
        "200":
          description: "Consulta de formas de pagamento foi bem sucedida"
          schema:
            type: "array"
            description: "Lista das formas de pagamento"
            items:
              $ref: "#/definitions/inline_response_200_16"
      x-swagger-router-controller: "Payment"
...
  /v1/setting/:
    get:
      tags:
      - "Setting"
      summary: "Configuração do backoffice"
      description: "Consulta a configuração do backoffice"
      operationId: "services.setting.web.get_setting"
      parameters:
      - name: "Authorization"
        in: "header"
        description: "Token de autorização da API gerado pelo hub de serviços (Em\
          \ rede local gerado pelo backoffice)"
        required: true
        type: "string"
      - name: "Store-CNPJ"
        in: "header"
        description: "CNPJ da empresa (somente digitos)"
        required: true
        type: "string"
      - name: "APK-Version"
        in: "header"
        description: "Número da versão do aplicativo"
        required: false
        type: "string"
      - name: "Device-Number"
        in: "header"
        description: "Número serial do dispositivo"
        required: true
        type: "string"
      - name: "RFID-Available"
        in: "header"
        description: "Indica se o dispositivo faz leitura de cartão RFID"
        required: false
        type: "boolean"
      responses:
        "200":
          description: "Consulta da configuração foi bem sucedida"
          schema:
            $ref: "#/definitions/inline_response_200_26"
      x-swagger-router-controller: "Setting"
definitions:
  Error:
    type: "object"
    properties:
      detail:
        type: "string"
        description: "Descrição do erro"
      status:
        type: "integer"
        description: "Status code do erro"
      title:
        type: "string"
        description: "Título do erro"
      type:
        type: "string"
        description: "Tipo do erro ocorrido"
    description: "Dados do erro"
  data:
    type: "object"
    properties:
      username:
        type: "string"
        description: "Nome de usuário do parceiro"
      password:
        type: "string"
        description: "Senha de acesso do parceiro"
  inline_response_200:
    type: "object"
    properties:
      token:
        type: "string"
        description: "Token de acesso padrão JWT"
      expiresIn:
        type: "string"
        description: "Tempo em milesegundos de validade do token"
    example:
      expiresIn: "expiresIn"
      token: "token"
...
  inline_response_200_7:
    type: "object"
    properties:
      customers:
        type: "array"
        description: "Lista de clientes"
        items:
          $ref: "#/definitions/inline_response_200_7_customers"
      nextPage:
        type: "integer"
        description: "Número da próxima página (Nulo se não houver próxima página)"
      prevPage:
        type: "integer"
        description: "Número da página anterior (Nulo se não houver página anterior)"
    example:
      nextPage: 1
      prevPage: 5
      customers:
      - code: 6
        driver:
          requestPassword: true
          name: "name"
        name: "name"
        nickname: "nickname"
        cpf: "cpf"
        plate: "plate"
        id: 0
        loyaltyCard: "loyaltyCard"
      - code: 6
        driver:
          requestPassword: true
          name: "name"
        name: "name"
        nickname: "nickname"
        cpf: "cpf"
        plate: "plate"
        id: 0
        loyaltyCard: "loyaltyCard"
...
  inline_response_200_8:
    type: "object"
    properties:
      statusCode:
        type: "integer"
        description: "Código de status da consulta (1-Com restrição 2-Sem restriçã\
          o 3-Forma de pagamento não permitida)"
      statusMessage:
        type: "string"
        description: "Mensagem de status da consulta"
      restrictions:
        type: "array"
        description: "Lista de restrições de crédito"
        items:
          type: "string"
          description: "Descrição da restrição de crédito"
      details:
        $ref: "#/definitions/inline_response_200_8_details"
    example:
      restrictions:
      - "restrictions"
      - "restrictions"
      details:
        summary:
        - totalAmount: 5.02500479152029466689555192715488374233245849609375
          amountBilled: 7.4577447736837658709418974467553198337554931640625
          amountAvailable: 6.8468526983526398765889098285697400569915771484375
          amountPending: 1.173074250955943309548956676735542714595794677734375
          amountUsed: 4.9652184929849543237878606305457651615142822265625
          paymentName: "paymentName"
        - totalAmount: 5.02500479152029466689555192715488374233245849609375
          amountBilled: 7.4577447736837658709418974467553198337554931640625
          amountAvailable: 6.8468526983526398765889098285697400569915771484375
          amountPending: 1.173074250955943309548956676735542714595794677734375
          amountUsed: 4.9652184929849543237878606305457651615142822265625
          paymentName: "paymentName"
        drivers:
        - blocked: true
          name: "name"
          id: 3
          blockedWeekend: true
          blockingReason: "blockingReason"
          limits:
          - totalAmount: 7.3862819483858839220147274318151175975799560546875
            amountAvailable: 2.027123023002321833274663731572218239307403564453125
            amountUsed: 4.1456080298839363962315474054776132106781005859375
            typeName: "typeName"
            litersAvailable: 1.231513536777255612975068288506008684635162353515625
            litersUsed: 1.024645700144157789424070870154537260532379150390625
            totalLiters: 1.489415909854170383397331534069962799549102783203125
          - totalAmount: 7.3862819483858839220147274318151175975799560546875
            amountAvailable: 2.027123023002321833274663731572218239307403564453125
            amountUsed: 4.1456080298839363962315474054776132106781005859375
            typeName: "typeName"
            litersAvailable: 1.231513536777255612975068288506008684635162353515625
            litersUsed: 1.024645700144157789424070870154537260532379150390625
            totalLiters: 1.489415909854170383397331534069962799549102783203125
        - blocked: true
          name: "name"
          id: 3
          blockedWeekend: true
          blockingReason: "blockingReason"
          limits:
          - totalAmount: 7.3862819483858839220147274318151175975799560546875
            amountAvailable: 2.027123023002321833274663731572218239307403564453125
            amountUsed: 4.1456080298839363962315474054776132106781005859375
            typeName: "typeName"
            litersAvailable: 1.231513536777255612975068288506008684635162353515625
            litersUsed: 1.024645700144157789424070870154537260532379150390625
            totalLiters: 1.489415909854170383397331534069962799549102783203125
          - totalAmount: 7.3862819483858839220147274318151175975799560546875
            amountAvailable: 2.027123023002321833274663731572218239307403564453125
            amountUsed: 4.1456080298839363962315474054776132106781005859375
            typeName: "typeName"
            litersAvailable: 1.231513536777255612975068288506008684635162353515625
            litersUsed: 1.024645700144157789424070870154537260532379150390625
            totalLiters: 1.489415909854170383397331534069962799549102783203125
        account:
          totalAmount: 5.962133916683182377482808078639209270477294921875
          amountAvailable: 6.02745618307040320615897144307382404804229736328125
          amountUsed: 1.46581298050294517310021547018550336360931396484375
        customer:
          totalAmount: 9.301444243932575517419536481611430644989013671875
          amountAvailable: 2.3021358869347654518833223846741020679473876953125
          amountUsed: 7.061401241503109105224211816675961017608642578125
          name: "name"
          id: 5
      statusMessage: "statusMessage"
      statusCode: 0
  inline_response_200_9:
    type: "object"
    properties:
      name:
        type: "string"
        description: "Nome do motorista"
      cpf:
        type: "string"
        description: "CPF formatado do motorista"
      requestPassword:
        type: "boolean"
        description: "Indica se é obrigatória informar a senha do motorista"
    description: "Dados do motorista"
    example:
      requestPassword: true
      name: "name"
      cpf: "cpf"
  inline_response_200_10:
    type: "object"
    properties:
      plate:
        type: "string"
        description: "Placa do veículo"
      vehicle:
        type: "string"
        description: "Descrição do veículo"
    description: "Dados do veículo da frota"
    example:
      plate: "plate"
      vehicle: "vehicle"
...
  inline_response_200_11:
    type: "object"
    properties:
      employees:
        type: "array"
        items:
          $ref: "#/definitions/inline_response_200_11_employees"
      nextPage:
        type: "integer"
        description: "Número da próxima página (nulo se não houver próxima página)"
      prevPage:
        type: "integer"
        description: "Número da página anterior (nulo se não houver página anterior)"
    description: "Lista de funcionários"
    example:
      nextPage: 1
      prevPage: 5
      employees:
      - code: 6
        name: "name"
        rfid: "rfid"
        id: 0
      - code: 6
        name: "name"
        rfid: "rfid"
        id: 0
...
  inline_response_200_16:
    type: "object"
    properties:
      dtef:
        type: "integer"
        description: "1-Crédito 2-Débito 4-Voucher (Conforme DTef)"
      id:
        type: "integer"
        description: "Id da forma de pagamento (1-Dinheiro 2-Crédito 3-Débito 4-Faturado\
          \ 5-ShellBox 6-Voucher 7-QRLinx PIX)"
      name:
        type: "string"
        description: "Nome da forma de pagamento"
      creditCheck:
        type: "boolean"
        description: "Indicador de consulta de restrição de crédito obrigatória para\
          \ autorizar a forma de pagamento na venda"
    description: "Dados da forma de pagamento"
    example:
      creditCheck: true
      name: "name"
      id: 6
      dtef: 0
...
  inline_response_200_26:
    type: "object"
    properties:
      showProducts:
        type: "boolean"
        description: "Indica se deve exibir os produtos"
      showFuelPumps:
        type: "boolean"
        description: "Indica se deve exibir os bicos de abastecimento"
      requestSalesman:
        type: "boolean"
        description: "Indica se deve solicitar o vendedor"
      requestLogin:
        type: "boolean"
        description: "Indica se deve solicitar o login"
      checkItemInventory:
        type: "boolean"
        description: "Indica se deve validar estoque ao lançar o item no carrinho"
      voucherPostoAki:
        type: "boolean"
        description: "Indica se deve solicitar o código do voucher PostoAki"
      useRfidCard:
        type: "integer"
        description: "Modo cartão RFID (0-Inativo 1-Ativo 2-Ativo e obrigatório)"
      groupedUserMode:
        type: "integer"
        description: "Modo de agrupamento por usuário (0-Não agrupado 1-Agrupado 2-Agrupado\
          \ com senha)"
      shiftWorkMode:
        type: "integer"
        description: "Modo de abertura/fechamento do turno (1-Automático 2-Manual\
          \ pelo PDV 3-Manual pelo POS)"
      closeReportPrintMode:
        type: "integer"
        description: "Modo de impressão do relatório de fechamento (1-Sem impressã\
          o 2-Questiona impressão 3-Impressão automática)"
      isPafEcf:
        type: "boolean"
        description: "Indica se deve operar com PAF-ECF (Programa Aplicativo Fiscal\
          \ com Emissor de Cupom Fiscal)"
      codeActive:
        type: "string"
        description: "Código de Ativação (Código para ativar SDK Stone)"
      setLayoutColumns:
        type: "boolean"
        description: "Indica se pode especificar a quantidade de colunas para impressã\
          o"
      loyalty:
        type: "boolean"
        description: "Indica se o programa de fidelidade para pontuação está ativado"
      replaceNotApprovedPayment:
        type: "boolean"
        description: "Indica se pode substituir os pagamentos não aprovados (Sempre\
          \ enviar True)"
      backoffice:
        $ref: "#/definitions/inline_response_200_26_backoffice"
      franchise:
        $ref: "#/definitions/inline_response_200_26_franchise"
      payments:
        type: "array"
        description: "Lista de formas de pagamento"
        items:
          $ref: "#/definitions/inline_response_200_16"
      withdrawals:
        type: "array"
        description: "Lista de formas de sangria"
        items:
          $ref: "#/definitions/inline_response_200_26_withdrawals"
      voucherSuperApp:
        type: "object"
        description: "Configuração do voucher SuperAPP"
        properties: {}
    example:
      codeActive: "codeActive"
      showProducts: true
      franchise:
        showLoyaltyLogo: true
        name: "name"
        id: 2
      closeReportPrintMode: 5
      checkItemInventory: true
      loyalty: true
      payments:
      - creditCheck: true
        name: "name"
        id: 6
        dtef: 0
      - creditCheck: true
        name: "name"
        id: 6
        dtef: 0
      groupedUserMode: 6
      showFuelPumps: true
      useRfidCard: 0
      isPafEcf: true
      withdrawals:
      - name: "name"
        id: 7
      - name: "name"
        id: 7
      voucherSuperApp: "{}"
      setLayoutColumns: true
      shiftWorkMode: 1
      backoffice:
        phone: "phone"
        name: "name"
        id: 5
      requestSalesman: true
      requestLogin: true
      voucherPostoAki: true
      replaceNotApprovedPayment: true
...